Fixes #10 and #11
Checked on Windows 7, 8.1 and 10 OK.
The actual error is being masked by built in WindowsError exception
being fetched from the _winreg module. This is now fixed, but the
underlying problem remains.
Please check the registry setting for
HKEY_CURRENT_USER\Software\node.js\InstallPath exists and that no
restrictions on the computer prevent access to the registry.